requirements.txt
akad==0.11.1
asgiref==1.1.2
astroid==1.6.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
beautifulsoup4==4.6.0
certifi==2017.11.5
channels==1.1.8
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
curve==0.1.0
daphne==1.4.2
Django==2.0.1
django-controlcenter==0.2.6
django-crispy-forms==1.6.1
django-extensions==2.0.7
django-filter==1.1.0
django-material==1.2.5
django-pkgconf==0.3.0
django-registration-redux==2.4
djangorestframework==3.8.2
ez-setup==0.9
google==2.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.2.15
lazy-object-proxy==1.3.1
line==0.8.2
linepy==3.0.8
lxml==4.2.1
Markdown==2.6.11
mccabe==0.6.1
Pillow==5.1.0
pyasn1==0.4.2
pylint==1.8.1
PyQRCode==1.2.1
pytz==2017.3
recursive==1.3.0
requests==2.18.4
rsa==3.4.2
six==1.11.0
thrift==0.11.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
wrapt==1.10.11
zope.interface==4.5.0
